requirements.txt
face_recognition
pillow
openpyxl
tqdm